HOTT-2132 Namespaced news models
