+++ app/controllers/feed/news_items_controller.rb
@@ -1,7 +1,7 @@
 module Feed
   class NewsItemsController < ApplicationController
     def index
-      @news_items = NewsItem.for_feed
+      @news_items = News::Item.for_feed
     end
   end
 end
