+++ app/controllers/find_commodities_controller.rb
@@ -3,6 +3,6 @@ class FindCommoditiesController < ApplicationController
 
   def show
     @no_shared_search = true
-    @latest_news = NewsItem.latest_for_home_page
+    @latest_news = News::Item.latest_for_home_page
   end
 end
