+++ app/models/news/collection.rb
@@ -1,0 +1,9 @@
+require 'api_entity'
+
+module News
+  class Collection
+    include ApiEntity
+
+    attr_accessor :name
+  end
+end
