--- app/models/news_collection.rb
@@ -1,7 +1,0 @@
-require 'api_entity'
-
-class NewsCollection
-  include ApiEntity
-
-  attr_accessor :name
-end
