+++ spec/factories/news/collection_factory.rb
@@ -1,5 +1,5 @@
 FactoryBot.define do
-  factory :news_collection do
+  factory :news_collection, class: 'News::Collection' do
     sequence(:id) { |n| n }
     sequence(:name) { |n| "Collection #{n}" }
   end
