+++ spec/features/errors_spec.rb
@@ -92,7 +92,7 @@ RSpec.describe 'Error handling' do
 
     context 'with non existant resource' do
       before do
-        stub_api_request('/news_items/9999', backend: 'uk')
+        stub_api_request('/news/items/9999', backend: 'uk')
           .and_return jsonapi_error_response(404)
 
         visit '/news/9999'
@@ -103,7 +103,7 @@ RSpec.describe 'Error handling' do
 
     context 'with faraday connection error' do
       before do
-        stub_api_request('/news_items/9999', backend: 'uk').to_timeout
+        stub_api_request('/news/items/9999', backend: 'uk').to_timeout
 
         visit '/news/9999'
       end
@@ -113,7 +113,7 @@ RSpec.describe 'Error handling' do
 
     context 'with other error' do
       before do
-        allow(NewsItem).to \
+        allow(News::Item).to \
           receive(:find).and_raise(StandardError, 'Something went wrong')
 
         visit '/news/9999'
@@ -124,7 +124,7 @@ RSpec.describe 'Error handling' do
 
     context 'with feature that is unavailable' do
       before do
-        allow(NewsItem).to \
+        allow(News::Item).to \
           receive(:find).and_raise(TradeTariffFrontend::FeatureUnavailable)
 
         visit '/news/9999'
