+++ spec/models/news/collection_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
 
-RSpec.describe NewsCollection do
+RSpec.describe News::Collection do
   it { is_expected.to respond_to :name }
 end
