+++ spec/models/news/item_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-RSpec.describe NewsItem do
+RSpec.describe News::Item do
   # This is stubbed for _every_ spec because its in the header menu
   # Reverting it to real implementation for this spec
   before do
@@ -29,7 +29,7 @@ RSpec.describe NewsItem do
       include_context 'with UK service'
 
       before do
-        stub_api_request('/news_items?service=uk&target=home&per_page=1', backend: 'uk')
+        stub_api_request('/news/items?service=uk&target=home&per_page=1', backend: 'uk')
           .to_return jsonapi_response :news_item, attributes_for_list(:news_item, 1)
       end
 
@@ -40,7 +40,7 @@ RSpec.describe NewsItem do
       include_context 'with XI service'
 
       before do
-        stub_api_request('/news_items?service=xi&target=home&per_page=1', backend: 'uk')
+        stub_api_request('/news/items?service=xi&target=home&per_page=1', backend: 'uk')
           .to_return jsonapi_response :news_item, attributes_for_list(:news_item, 1)
       end
 
@@ -55,7 +55,7 @@ RSpec.describe NewsItem do
       include_context 'with UK service'
 
       before do
-        stub_api_request('/news_items?service=uk&target=banner&per_page=1', backend: 'uk')
+        stub_api_request('/news/items?service=uk&target=banner&per_page=1', backend: 'uk')
           .to_return jsonapi_response :news_item, attributes_for_list(:news_item, 1, :banner)
       end
 
@@ -66,7 +66,7 @@ RSpec.describe NewsItem do
       include_context 'with XI service'
 
       before do
-        stub_api_request('/news_items?service=xi&target=banner&per_page=1', backend: 'uk')
+        stub_api_request('/news/items?service=xi&target=banner&per_page=1', backend: 'uk')
           .to_return jsonapi_response :news_item, attributes_for_list(:news_item, 1, :banner)
       end
 
@@ -75,7 +75,7 @@ RSpec.describe NewsItem do
 
     context 'with failed connection to backend' do
       before do
-        stub_api_request('/news_items?service=uk&target=banner&per_page=1', backend: 'uk')
+        stub_api_request('/news/items?service=uk&target=banner&per_page=1', backend: 'uk')
           .to_timeout
       end
 
@@ -90,7 +90,7 @@ RSpec.describe NewsItem do
       include_context 'with UK service'
 
       before do
-        stub_api_request('/news_items?page=1&service=uk&target=updates', backend: 'uk')
+        stub_api_request('/news/items?page=1&service=uk&target=updates', backend: 'uk')
           .to_return jsonapi_response :news_item, attributes_for_list(:news_item, 2)
       end
 
@@ -101,7 +101,7 @@ RSpec.describe NewsItem do
       include_context 'with XI service'
 
       before do
-        stub_api_request('/news_items?page=1&service=xi&target=updates', backend: 'uk')
+        stub_api_request('/news/items?page=1&service=xi&target=updates', backend: 'uk')
           .to_return jsonapi_response :news_item, attributes_for_list(:news_item, 2)
       end
 
@@ -112,7 +112,7 @@ RSpec.describe NewsItem do
       subject { described_class.updates_page(2) }
 
       before do
-        stub_api_request('/news_items?page=2&service=uk&target=updates', backend: 'uk')
+        stub_api_request('/news/items?page=2&service=uk&target=updates', backend: 'uk')
           .to_return jsonapi_response :news_item, attributes_for_list(:news_item, 2)
       end
 
@@ -126,7 +126,7 @@ RSpec.describe NewsItem do
     include_context 'with UK service'
 
     before do
-      stub_api_request('/news_items?target=updates', backend: 'uk')
+      stub_api_request('/news/items?target=updates', backend: 'uk')
         .to_return jsonapi_response :news_item, attributes_for_list(:news_item, 2)
     end
 
@@ -142,7 +142,7 @@ RSpec.describe NewsItem do
       include_context 'with UK service'
 
       before do
-        stub_api_request('/news_items?page=1&service=uk&target=updates', backend: 'uk')
+        stub_api_request('/news/items?page=1&service=uk&target=updates', backend: 'uk')
           .to_return jsonapi_response :news_item, updates
       end
 
@@ -169,7 +169,7 @@ RSpec.describe NewsItem do
       include_context 'with XI service'
 
       before do
-        stub_api_request('/news_items?page=1&service=xi&target=updates', backend: 'uk')
+        stub_api_request('/news/items?page=1&service=xi&target=updates', backend: 'uk')
           .to_return jsonapi_response :news_item, updates
       end
 
@@ -194,7 +194,7 @@ RSpec.describe NewsItem do
 
     context 'with failed connection to backend' do
       before do
-        stub_api_request('/news_items?page=1&service=uk&target=updates', backend: 'uk')
+        stub_api_request('/news/items?page=1&service=uk&target=updates', backend: 'uk')
           .to_timeout
       end
 
@@ -211,7 +211,7 @@ RSpec.describe NewsItem do
       include_context 'with UK service'
 
       before do
-        stub_api_request('/news_items?service=uk&target=banner&per_page=1', backend: 'uk')
+        stub_api_request('/news/items?service=uk&target=banner&per_page=1', backend: 'uk')
           .to_return jsonapi_response :news_item, attributes_for_list(:news_item, 1, :banner)
       end
 
@@ -228,7 +228,7 @@ RSpec.describe NewsItem do
       include_context 'with XI service'
 
       before do
-        stub_api_request('/news_items?service=xi&target=banner&per_page=1', backend: 'uk')
+        stub_api_request('/news/items?service=xi&target=banner&per_page=1', backend: 'uk')
           .to_return jsonapi_response :news_item, attributes_for_list(:news_item, 1, :banner)
       end
 
@@ -300,7 +300,7 @@ RSpec.describe NewsItem do
     subject { build(:news_item, collection_count: 2).collections }
 
     it { is_expected.to have_attributes length: 2 }
-    it { is_expected.to all be_instance_of NewsCollection }
+    it { is_expected.to all be_instance_of News::Collection }
   end
 
   describe '#precis' do
