+++ spec/spec_helper.rb
@@ -61,8 +61,8 @@ RSpec.configure do |config|
 
   config.before do
     allow(TariffUpdate).to receive(:all).and_return([OpenStruct.new(applied_at: Time.zone.today)])
-    allow(NewsItem).to receive(:any_updates?).and_return true
-    allow(NewsItem).to receive(:latest_banner).and_return build(:news_item, :banner)
+    allow(News::Item).to receive(:any_updates?).and_return true
+    allow(News::Item).to receive(:latest_banner).and_return build(:news_item, :banner)
     Thread.current[:service_choice] = nil
   end
