+++ spec/support/shared_context/latest_news.rb
@@ -1,5 +1,5 @@
 RSpec.shared_context 'with latest news stubbed' do
-  before { allow(NewsItem).to receive(:latest_for_home_page).and_return latest_news }
+  before { allow(News::Item).to receive(:latest_for_home_page).and_return latest_news }
 
   let(:latest_news) { build :news_item }
 end
