+++ spec/vcr/news_items_updates.yml
@@ -2,7 +2,7 @@
 http_interactions:
 - request:
     method: get
-    uri: http://localhost:3018/news_items?target=updates
+    uri: http://localhost:3018/news/items?target=updates
     body:
       encoding: US-ASCII
       string: ''
